finished

+++ src/app/componentes/task-component/task-component.component.ts
@@ -4,27 +4,24 @@ import { Task } from '../../modelos/task';
 
 import { faTrashAlt } from '@fortawesome/free-solid-svg-icons';
 import { TaskService } from '../../servicios/task-service.service';
-import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-task-component',
   templateUrl: './task-component.component.html',
-  styleUrls: ['./task-component.component.css']
+  styleUrls: ['./task-component.component.css'],
 })
 export class TaskComponentComponent implements OnInit {
-
   @Input() task!: Task;
 
   faTrashAlt = faTrashAlt;
 
-  constructor(public taskService: TaskService, private toastr: ToastrService) { }
+  constructor(public taskService: TaskService) {}
 
-  ngOnInit(): void {
-  }
+  ngOnInit(): void {}
 
   deleteTask(task: Task) {
-    this.taskService.deleteTask(task);
-    this.toastr.success('Task deleted successfully.', 'Success!');
+    if (confirm('¿Estás seguro de que quieres eliminar esta tarea?')) {
+      this.taskService.deleteTask(task);
+    }
   }
-
 }
